fix: WebAssembly Compilation (#1)
* feat: conditional compilation for wasm u64 to u32 casting

* Add instructions and fix build and component

* fix readme

* fix formatting?

* fix more formatting

* fix formatting again

* fix typo

* cast as size_t instead

+++ src/world.rs
@@ -297,8 +297,9 @@ impl Drop for World {
 // Additional Add-ons support
 impl World {
 	pub fn enable_rest(&self) {
-		let rest_comp_id = unsafe { FLECS__EEcsRest as u64 };
-		let rest_comp_size = std::mem::size_of::<EcsRest>() as u64;
+    let rest_comp_id = unsafe { FLECS__EEcsRest as u64 };
+		let rest_comp_size = std::mem::size_of::<EcsRest>() as size_t;
+		
 		let rest_data: EcsRest = unsafe { MaybeUninit::zeroed().assume_init() };
 
 		unsafe {
